% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taiwan_aq.R
\name{download_parquet}
\alias{download_parquet}
\title{Download the parquet file for a given year and item, if not already cached}
\usage{
download_parquet(year, item, version = "v0.1.0")
}
\description{
Download the parquet file for a given year and item, if not already cached
}
